Moving view warming into its own file
